rename sticky header to global header
