+++ app/application/libraries/ExploreUK/views/global-header.php
