+++ app/application/libraries/ExploreUK/views/header.php
@@ -45,13 +45,13 @@
 <?php if ($m['front_page']) : ?>
     <section id="top">
         <div>
-    <?php require('stickyheader.php'); ?>
+    <?php require('global-header.php'); ?>
     <?php require('search.php'); ?>
 <div class="row"></div>
 <?php else : ?>
     <section class="bg-uklblack" id="top2">
         <div>
-    <?php require('stickyheader.php'); ?>
+    <?php require('global-header.php'); ?>
 <div id="top2_bottom">
     <?php require('search-brief.php'); ?>
         </div>
